+++ src/clj/ebtanas/core.clj
@@ -1,6 +1,6 @@
 (ns ebtanas.core
   (:require [ring.adapter.jetty :as jetty]
-            [ring.middleware.defaults :refer [wrap-defaults site-defaults]]
+            [ring.middleware.defaults :refer [wrap-defaults site-defaults api-defaults]]
             [ring.middleware.reload :refer [wrap-reload]]
             [ring.middleware.gzip :refer [wrap-gzip]]
             [ring.middleware.webjars :refer [wrap-webjars]]
@@ -8,12 +8,14 @@
             [compojure.route :refer [resources not-found]]
             [ebtanas.routes.public :refer [public-routes]]
             [ebtanas.routes.backend :refer [backend-routes]]))
+            ;;[ebtanas.routes.services :refer [service-route]]))
 
 (def all-routes
   (routes
     (resources "/")
     public-routes
     backend-routes
+    ;;service-route
     (not-found "<h1>404</h1>")))
 
 (defonce server
@@ -22,7 +24,7 @@
       (-> #'all-routes
           (wrap-reload)
           (wrap-webjars)
-          (wrap-defaults site-defaults)
+          (wrap-defaults site-defaults) ;;api-defaults)
           (wrap-gzip))
       {:port (if port (Integer. port) 3000)
        :join false})))